improve

+++ router_compile.sh
@@ -7,14 +7,18 @@ lede_dir=$(pwd)
 cp $scripts_dir/conf/Router.buildinfo $lede_dir/.config
 # 开始编译
 $scripts_dir/compile.sh
-# 时间
-date=$(date +"%Y-%m-%d_%H:%M:%S")
-# 整理文件
-rm -rf $(find $lede_dir/bin/targets/ -type d -name "packages")
-# 创建目录
-mkdir -vp $lede_dir/Firmware/$date/Router/{firmware,buildinfo}/
-# 移动固件文件
-mv -f $(find $lede_dir/bin/targets/ -type f) $lede_dir/Firmware/$date/Router/firmware/
-mv -f $(find $lede_dir/bin/targets/ -type f -name "*.buildinfo" -o -name "*.manifest") $lede_dir/Firmware/$date/Router/buildinfo/
-# 展示目录
-tree $lede_dir/Firmware/$date/Router
+if [ ! -e $lede_dir/bin/targets/**/**/sha256sums ]
+then
+    # 时间
+    date=$(date +"%Y-%m-%d_%H:%M:%S")
+    # 整理文件
+    rm -rf $(find $lede_dir/bin/targets/ -type d -name "packages")
+    # 创建目录
+    mkdir -vp $lede_dir/Firmware/$date/Router/{firmware,buildinfo}/
+    # 移动固件文件
+    echo '\033[32m moving firmware file \033[0m'
+    mv -f $(find $lede_dir/bin/targets/ -type f) $lede_dir/Firmware/$date/Router/firmware/
+    mv -f $(find $lede_dir/bin/targets/ -type f -name "*.buildinfo" -o -name "*.manifest") $lede_dir/Firmware/$date/Router/buildinfo/
+    # 展示目录
+    tree $lede_dir/Firmware/$date/Router
+fi
